backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
langchain
langchain-openai
pandas
numpy
plotly
python-dateutil
pydantic
groq
langchain-groq